+++ bill/invoice.go
@@ -146,6 +146,7 @@ func (inv *Invoice) ValidateWithContext(ctx context.Context) error {
 		validation.Field(&inv.ValueDate),
 		validation.Field(&inv.Currency,
 			validation.Required,
+			currency.CanExchangeTo(inv.ExchangeRates, r.Currency),
 		),
 		validation.Field(&inv.ExchangeRates),
 		validation.Field(&inv.Preceding),
@@ -438,14 +439,10 @@ func (inv *Invoice) calculateWithRegime(r *tax.Regime) error {
 	t.reset(zero)
 
 	// Lines
-	if err := calculateLines(r, inv.Lines, inv.Currency); err != nil {
+	if err := calculateLines(r, inv.Lines, inv.Currency, inv.ExchangeRates); err != nil {
 		return validation.Errors{"lines": err}
 	}
-	var err error
-	t.Sum, err = calculateLineSum(inv.Lines, inv.Currency, inv.ExchangeRates)
-	if err != nil {
-		return validation.Errors{"lines": err}
-	}
+	t.Sum = calculateLineSum(inv.Lines, inv.Currency)
 	t.Total = t.Sum
 
 	// Discount Lines