+++ currency/currency.go
@@ -14,5 +14,6 @@ func init() {
 	schema.Register(schema.GOBL.Add("currency"),
 		Code(""),
 		ExchangeRate{},
+		Amount{},
 	)
 }