+++ currency/exchange_rate_test.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/invopop/gobl/currency"
 	"github.com/invopop/gobl/num"
+	"github.com/invopop/validation"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -18,7 +19,7 @@ func TestExchangeRateValidation(t *testing.T) {
 			name: "valid",
 			rate: currency.ExchangeRate{
 				From:   currency.USD,
-				Into:   currency.EUR,
+				To:     currency.EUR,
 				Amount: num.MakeAmount(875967, 6),
 			},
 			exp: "",
@@ -28,15 +29,24 @@ func TestExchangeRateValidation(t *testing.T) {
 			rate: currency.ExchangeRate{
 				Amount: num.MakeAmount(875967, 6),
 			},
-			exp: "from: cannot be blank, into: cannot be blank",
+			exp: "from: cannot be blank; to: cannot be blank",
 		},
 		{
 			name: "missing amount",
 			rate: currency.ExchangeRate{
 				From: currency.USD,
-				Into: currency.EUR,
+				To:   currency.EUR,
 			},
-			exp: "amount: must not be zero",
+			exp: "amount: must be greater than 0",
+		},
+		{
+			name: "negative amount",
+			rate: currency.ExchangeRate{
+				From:   currency.USD,
+				To:     currency.EUR,
+				Amount: num.MakeAmount(-87596, 3),
+			},
+			exp: "amount: must be greater than 0",
 		},
 	}
 	for _, test := range tests {
@@ -52,18 +62,7 @@ func TestExchangeRateValidation(t *testing.T) {
 }
 
 func TestMatchExchangeRate(t *testing.T) {
-	rates := []*currency.ExchangeRate{
-		{
-			From:   currency.USD,
-			Into:   currency.EUR,
-			Amount: num.MakeAmount(875967, 6),
-		},
-		{
-			From:   currency.EUR,
-			Into:   currency.USD,
-			Amount: num.MakeAmount(1141860, 6),
-		},
-	}
+	rates := sampleRates()
 	a := currency.MatchExchangeRate(rates, currency.USD, currency.EUR)
 	assert.Equal(t, "0.875967", a.String())
 
@@ -78,23 +77,7 @@ func TestMatchExchangeRate(t *testing.T) {
 }
 
 func TestExchange(t *testing.T) {
-	rates := []*currency.ExchangeRate{
-		{
-			From:   currency.USD,
-			Into:   currency.EUR,
-			Amount: num.MakeAmount(875967, 6),
-		},
-		{
-			From:   currency.EUR,
-			Into:   currency.USD,
-			Amount: num.MakeAmount(1141860, 6),
-		},
-		{
-			From:   currency.EUR,
-			Into:   currency.CLP,
-			Amount: num.MakeAmount(100629, 2),
-		},
-	}
+	rates := sampleRates()
 	a := num.MakeAmount(10000, 2)
 	b := currency.Exchange(rates, currency.USD, currency.EUR, a)
 	assert.Equal(t, "87.60", b.String())
@@ -110,4 +93,40 @@ func TestExchange(t *testing.T) {
 
 	b = currency.Exchange(rates, currency.EUR, currency.CLP, a)
 	assert.Equal(t, "100629", b.String())
+}
+
+func TestExchangeRateValidationRule(t *testing.T) {
+	rates := sampleRates()
+	cur := currency.USD
+	err := validation.Validate(cur, currency.CanExchangeTo(rates, currency.EUR))
+	assert.NoError(t, err)
+
+	err = validation.Validate(cur, currency.CanExchangeTo(rates, currency.MXN))
+	assert.ErrorContains(t, err, "no exchange rate defined for 'USD' to 'MXN")
+
+	err = validation.Validate(currency.CodeEmpty, currency.CanExchangeTo(rates, currency.EUR))
+	assert.NoError(t, err)
+
+	err = validation.Validate(currency.CodeEmpty, currency.CanExchangeTo(rates, currency.USD))
+	assert.NoError(t, err)
+}
+
+func sampleRates() []*currency.ExchangeRate {
+	return []*currency.ExchangeRate{
+		{
+			From:   currency.USD,
+			To:     currency.EUR,
+			Amount: num.MakeAmount(875967, 6),
+		},
+		{
+			From:   currency.EUR,
+			To:     currency.USD,
+			Amount: num.MakeAmount(1141860, 6),
+		},
+		{
+			From:   currency.EUR,
+			To:     currency.CLP,
+			Amount: num.MakeAmount(100629, 2),
+		},
+	}
 }