+++ regimes/co/invoices.go
@@ -2,7 +2,6 @@ package co
 
 import (
 	"github.com/invopop/gobl/bill"
-	"github.com/invopop/gobl/currency"
 	"github.com/invopop/gobl/l10n"
 	"github.com/invopop/gobl/org"
 	"github.com/invopop/gobl/tax"
@@ -21,7 +20,6 @@ func validateInvoice(inv *bill.Invoice) error {
 func (v *invoiceValidator) validate() error {
 	inv := v.inv
 	return validation.ValidateStruct(inv,
-		validation.Field(&inv.Currency, validation.In(currency.COP)),
 		validation.Field(&inv.Type,
 			validation.In(
 				bill.InvoiceTypeStandard,