+++ regimes/it/invoice_validator.go
@@ -5,7 +5,6 @@ import (
 	"regexp"
 
 	"github.com/invopop/gobl/bill"
-	"github.com/invopop/gobl/currency"
 	"github.com/invopop/gobl/l10n"
 	"github.com/invopop/gobl/org"
 	"github.com/invopop/gobl/tax"
@@ -30,9 +29,6 @@ func (v *invoiceValidator) validate() error {
 
 	inv := v.inv
 	return validation.ValidateStruct(inv,
-		validation.Field(&inv.Currency,
-			validation.In(currency.EUR),
-		),
 		validation.Field(&inv.Supplier,
 			validation.By(v.supplier),
 			validation.Skip,