+++ regimes/nl/invoices.go
@@ -2,7 +2,6 @@ package nl
 
 import (
 	"github.com/invopop/gobl/bill"
-	"github.com/invopop/gobl/currency"
 	"github.com/invopop/gobl/org"
 	"github.com/invopop/gobl/tax"
 	"github.com/invopop/validation"
@@ -22,9 +21,6 @@ func validateInvoice(inv *bill.Invoice) error {
 func (v *invoiceValidator) validate() error {
 	inv := v.inv
 	return validation.ValidateStruct(inv,
-		validation.Field(&inv.Currency,
-			validation.In(currency.EUR),
-		),
 		validation.Field(&inv.Supplier,
 			validation.By(v.supplier),
 		),