+++ regimes/pl/invoices.go
@@ -2,7 +2,6 @@ package pl
 
 import (
 	"github.com/invopop/gobl/bill"
-	"github.com/invopop/gobl/currency"
 	"github.com/invopop/gobl/org"
 	"github.com/invopop/gobl/tax"
 	"github.com/invopop/validation"
@@ -22,7 +21,6 @@ func validateInvoice(inv *bill.Invoice) error {
 func (v *invoiceValidator) validate() error {
 	inv := v.inv
 	return validation.ValidateStruct(inv,
-		validation.Field(&inv.Currency, validation.In(currency.PLN, currency.EUR)),
 		// Only commercial and simplified supported at this time for Poland.
 		// Rectification state determined by Preceding value.
 		validation.Field(&inv.Type, validation.In(