+++ regimes/us/invoices.go
@@ -2,7 +2,6 @@ package us
 
 import (
 	"github.com/invopop/gobl/bill"
-	"github.com/invopop/gobl/currency"
 	"github.com/invopop/validation"
 )
 
@@ -20,7 +19,6 @@ func validateInvoice(inv *bill.Invoice) error {
 func (v *invoiceValidator) validate() error {
 	inv := v.inv
 	return validation.ValidateStruct(inv,
-		validation.Field(&inv.Currency, validation.In(currency.USD)),
 		validation.Field(&inv.Supplier, validation.Required),
 	)
 }